backend/requirements.txt
fastapi
uvicorn
pymongo
httpx
python-dotenv
pydantic 
google-genai
